pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = [
    "hatchling",
]

[project]
name = "aa-wizard-skillfarm"
description = "Wizard skill farm"
readme = "README.md"
keywords = [
    "allianceauth",
    "wizard",
]
license = {file = "LICENSE"}
authors = [
    { name = "Dusty Meg", email = "aa_plugins@mmoore.uk" },
]
requires-python = ">=3.8"
classifiers = [
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]
dynamic = [
    "version",
]
dependencies = [
    "allianceauth<5.0.0,>=4",
    "allianceauth-corptools>=2.7",
]
[project.urls]
Changelog = "https://github.com/Dusty-Meg/aa-wizard-skillfarm/blob/master/CHANGELOG.md"
"Issue / Bug Reports" = "https://github.com/Dusty-Meg/aa-wizard-skillfarm/issues"

[tool.hatch.version]
path = "wizardskillfarm/__init__.py"

[tool.hatch.build]
include = [
    "/wizardskillfarm",
]
